suppress warnings and change some 1d arrays to std::vector

+++ src/ParameterReader.cpp
@@ -69,7 +69,7 @@ long ParameterReader::find(string name) {
   Check if the parameter with "name" already exists in the internal 
   "names" list. If yes, it returns its
 */
-    for (long ii=0; ii<names->size(); ii++)
+    for (unsigned int ii=0; ii<names->size(); ii++)
       if ((*names)[ii].compare(toLower(trim(name)))==0)
           return ii;
     return -1;
@@ -175,7 +175,7 @@ void ParameterReader::echo() {
     cout << "Parameter list: " << endl;
     cout << "=============================================================" 
          << endl;
-    for (long ii = 0; ii < names->size(); ii++)
+    for (unsigned int ii = 0; ii < names->size(); ii++)
         cout << (*names)[ii] << "=" << (*values)[ii] << endl;
     cout << "Parameter list end ==========================================" 
          << endl;
